# Optimize deletion